Fix invalid import
