+++ video/src/interactions.ts
@@ -1,0 +1,17 @@
+export enum VideoInteraction {
+    play = 'play',
+    restart = 'restart',
+    togglePlay = 'togglePlay',
+    pause = 'pause',
+    mute = 'mute',
+    unMute = 'unMute',
+    toggleSound = 'toggleSound',
+  }
+  
+  export enum VideoEvent {
+    onPlay = 'onPlay',
+    onPause = 'onPause',
+    onEnd = 'onEnd',
+    currentTime = 'currentTime',
+  }
+
